+++ src/app/profile/favorited-builds/actions.ts
@@ -26,6 +26,14 @@ import {
   archetypeFiltersToIds,
   limitByArchetypesSegment,
 } from '@/features/filters/queries/segments/limitByArchtypes'
+import {
+  amuletFilterToId,
+  limitByAmuletSegment,
+} from '@/features/filters/queries/segments/limitByAmulet'
+import {
+  limitByRingSegment,
+  ringFilterToId,
+} from '@/features/filters/queries/segments/limitByRing'
 
 export async function getFavoritedBuilds({
   communityBuildFilters,
@@ -43,23 +51,34 @@ export async function getFavoritedBuilds({
   const session = await getServerSession()
   const userId = session?.user?.id
 
-  const { archetypes, longGun, handGun, melee, selectedReleases } =
-    communityBuildFilters
-  const archetypeIds = archetypeFiltersToIds({ archetypes })
+  const {
+    archetypes,
+    longGun,
+    handGun,
+    melee,
+    ring,
+    amulet,
+    selectedReleases,
+  } = communityBuildFilters
 
   if (selectedReleases.length === 0) return { items: [], totalItemCount: 0 }
 
+  const archetypeIds = archetypeFiltersToIds({ archetypes })
   const weaponIds = weaponFiltersToIds({
     longGun,
     handGun,
     melee,
   })
+  const amuletId = amuletFilterToId({ amulet })
+  const ringId = ringFilterToId({ ring })
 
   const whereConditions = Prisma.sql`
 WHERE Build.isPublic = true
 AND Build.createdById != ${userId}
 ${limitByArchetypesSegment(archetypeIds)}
 ${limitByWeaponsSegment(weaponIds)}
+${limitByAmuletSegment(amuletId)}
+${limitByRingSegment(ringId)}
 ${limitByReleasesSegment(selectedReleases)}
 ${limitByTimeCondition(timeRange)}
 `