+++ src/features/filters/constants.ts
@@ -5,5 +5,7 @@ export const DEFAULT_COMMUNITY_BUILD_FILTERS: CommunityBuildFilterProps = {
   longGun: 'All',
   handGun: 'All',
   melee: 'All',
+  ring: 'All',
+  amulet: 'All',
   selectedReleases: ['base', 'dlc1'],
 }