+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## 2024-02-23
 
+### Added
+
+- You can now filter builds by an amulet or ring in the various build lists.
+
 ### Fixed
 
 - Fixed bug preventing duplication of builds